Modulo - Playing Now Media  - Refatorado: Criado algumas classes common para uso em diversos lugares sem necessidade de repetição de código.
