+++ app/src/main/java/ericambiel/com/br/smartimdb/ui/common/CommonMediaHolder.java
@@ -1,4 +1,4 @@
-package ericambiel.com.br.smartimdb.ui.util;
+package ericambiel.com.br.smartimdb.ui.common;
 
 
 import android.view.View;
