+++ app/src/main/java/ericambiel/com/br/smartimdb/ui/filmespopulares/FilmesPopularesContrato.kt
@@ -1,13 +1,13 @@
 package ericambiel.com.br.smartimdb.ui.filmespopulares
 
-import ericambiel.com.br.smartimdb.ui.MediaContrato
+import ericambiel.com.br.smartimdb.ui.common.CommonContrato
 
 interface FilmesPopularesContrato{
-    interface View : MediaContrato.View{
+    interface View : CommonContrato.View{
         //TODO
     }
 
-    interface Presenter : MediaContrato.Presenter{
+    interface Presenter : CommonContrato.Presenter{
         //TODO
     }
 }
