+++ app/src/main/res/layout/fragment_media.xml
@@ -4,7 +4,7 @@
     android:id="@+id/fl_media"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.youtubeplayer.YoutubeFragment">
+    tools:context=".ui.util.youtubeplayer.YoutubeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
